Related work items: #1116

+++ src/Stratis.Bitcoin.IntegrationTests/EnvironmentMockUpHelpers/NodeBuilder.cs
@@ -132,12 +132,11 @@ namespace Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers
             {
                 try
                 {
-                    Directory.Delete(folder, true);
-                    return true;
-                }
-                catch (DirectoryNotFoundException)
-                {
-                    return true;
+                    if (Directory.Exists(folder))
+                    {
+                        Directory.Delete(folder, true);
+                        return true;
+                    }
                 }
                 catch (Exception)
                 {
